move tests into lib.rs
